osint-platform-backend/requirements.txt
# OSINT Phone Intelligence Platform - Backend Requirements
# Python 3.9+

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
flask-talisman>=1.1.0

# Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# Data Processing
pydantic>=2.5.0
marshmallow>=3.20.0

# Phone Number Processing
phonenumbers>=8.13.32

# HTTP Requests & Web Scraping
requests>=2.31.0
requests-cache>=1.2.0
urllib3<2.0.0
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Search Engines
duckduckgo-search>=6.0.0
googlesearch-python>=1.2.3

# AI Services
# Note: For Gemini, install: pip install google-generativeai

# Database
sqlalchemy>=2.0.0
flask-sqlalchemy>=3.1.0

# Caching
redis>=5.0.0
diskcache>=5.6.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
colorama>=0.4.6
tqdm>=4.66.0

# Logging & Monitoring
structlog>=24.1.0
sentry-sdk[flask]>=1.40.0

# Testing
pytest>=8.0.0
pytest-flask>=1.3.0
pytest-cov>=4.1.0
